+++ common/src/main/java/net/anvian/bedrockplus/core/block/ModBlocks.java
@@ -1,5 +1,6 @@
 package net.anvian.bedrockplus.core.block;
 
+import net.anvian.bedrockplus.CommonMod;
 import net.anvian.bedrockplus.Constants;
 import net.anvian.bedrockplus.core.config.ModConfigs;
 import net.minecraft.core.registries.Registries;
@@ -12,11 +13,13 @@ import net.minecraft.world.level.block.state.BlockBehaviour;
 import net.minecraft.world.level.material.MapColor;
 
 public class ModBlocks {
+    private static final ModConfigs.BedrockPlusConfig config = CommonMod.configs.getConfig();
+
     public static final Block IMPURE_BEDROCK = new Block(
             BlockBehaviour.Properties
                     .ofFullCopy(Blocks.ANCIENT_DEBRIS)
-                    .strength((float) ModConfigs.bedrockImpureDeepslateHardness,
-                            (float) ModConfigs.bedrockImpureDeepslateResistance)
+                    .strength((float) config.bedrockImpureDeepslateHardness,
+                            (float) config.bedrockImpureDeepslateResistance)
                     .requiresCorrectToolForDrops()
                     .setId(key(Constants.IMPURE_BEDROCK_ID))
     );
@@ -24,8 +27,8 @@ public class ModBlocks {
     public static final Block IMPURE_BEDROCK_BLOCK = new Block(
             BlockBehaviour.Properties
                     .of().mapColor(MapColor.COLOR_BLACK)
-                    .strength((float) ModConfigs.blockOfImpureBedrockHardness,
-                            (float) ModConfigs.blockOfImpureBedrockResistance)
+                    .strength((float) config.blockOfImpureBedrockHardness,
+                            (float) config.blockOfImpureBedrockResistance)
                     .requiresCorrectToolForDrops().sound(SoundType.NETHERITE_BLOCK)
                     .setId(key(Constants.IMPURE_BEDROCK_BLOCK_ID))
     );