+++ common/src/main/java/net/anvian/bedrockplus/core/item/ModItems.java
@@ -1,5 +1,6 @@
 package net.anvian.bedrockplus.core.item;
 
+import net.anvian.bedrockplus.CommonMod;
 import net.anvian.bedrockplus.Constants;
 import net.anvian.bedrockplus.core.block.ModBlocks;
 import net.anvian.bedrockplus.core.config.ModConfigs;
@@ -17,6 +18,7 @@ import net.minecraft.world.item.ShovelItem;
 import net.minecraft.world.item.equipment.ArmorType;
 
 public class ModItems {
+    private static final ModConfigs.BedrockPlusConfig config = CommonMod.configs.getConfig();
     //blocks
     public static final BlockItem IMPURE_BEDROCK = new BlockItem(ModBlocks.IMPURE_BEDROCK, new Item.Properties().useBlockDescriptionPrefix().setId(key(Constants.IMPURE_BEDROCK_ID)));
     public static final BlockItem IMPURE_BEDROCK_BLOCK = new BlockItem(ModBlocks.IMPURE_BEDROCK_BLOCK, new Item.Properties().useBlockDescriptionPrefix().setId(key(Constants.IMPURE_BEDROCK_BLOCK_ID)));
@@ -27,25 +29,24 @@ public class ModItems {
     public static final Item IMPURE_BEDROCK_UPGRADE_SMITHING_TEMPLATE = ImpureBedrockSmithingTemplateItem.createImpureBedrockUpgradeTemplate(new Item.Properties().rarity(Rarity.UNCOMMON).setId(key(Constants.IMPURE_BEDROCK_UPGRADE_SMITHING_TEMPLATE_ID)));
 
     //tools
-
     public static final Item IMPUREBEDROCK_SWORD = new Item(
-            new Item.Properties().sword(ModMaterials.Tool.IMPURE_BEDROCK, ModConfigs.swordAttackDamage, (float) ModConfigs.swordAttackSpeed).fireResistant().setId(key(Constants.IMPUREBEDROCK_SWORD_ID))
+            new Item.Properties().sword(ModMaterials.Tool.IMPURE_BEDROCK, config.swordAttackDamage, (float) config.swordAttackSpeed).fireResistant().setId(key(Constants.IMPUREBEDROCK_SWORD_ID))
     );
 
     public static final Item IMPUREBEDROCK_PICKAXE = new Item(
-            new Item.Properties().pickaxe(ModMaterials.Tool.IMPURE_BEDROCK, (float) ModConfigs.pickaxeAttackDamage, (float) ModConfigs.pickaxeAttackSpeed).fireResistant().setId(key(Constants.IMPUREBEDROCK_PICKAXE_ID))
+            new Item.Properties().pickaxe(ModMaterials.Tool.IMPURE_BEDROCK, (float) config.pickaxeAttackDamage, (float) config.pickaxeAttackSpeed).fireResistant().setId(key(Constants.IMPUREBEDROCK_PICKAXE_ID))
     );
 
     public static final Item IMPUREBEDROCK_AXE = new AxeItem(
-            ModMaterials.Tool.IMPURE_BEDROCK, (float) ModConfigs.axeAttackDamage, (float) ModConfigs.axeAttackSpeed, new Item.Properties().fireResistant().setId(key(Constants.IMPUREBEDROCK_AXE_ID))
+            ModMaterials.Tool.IMPURE_BEDROCK, (float) config.axeAttackDamage, (float) config.axeAttackSpeed, new Item.Properties().fireResistant().setId(key(Constants.IMPUREBEDROCK_AXE_ID))
     );
 
     public static final Item IMPUREBEDROCK_SHOVEL = new ShovelItem(
-            ModMaterials.Tool.IMPURE_BEDROCK, (float) ModConfigs.shovelAttackDamage, (float) ModConfigs.shovelAttackSpeed, new Item.Properties().fireResistant().setId(key(Constants.IMPUREBEDROCK_SHOVEL_ID))
+            ModMaterials.Tool.IMPURE_BEDROCK, (float) config.shovelAttackDamage, (float) config.shovelAttackSpeed, new Item.Properties().fireResistant().setId(key(Constants.IMPUREBEDROCK_SHOVEL_ID))
     );
 
     public static final Item IMPUREBEDROCK_HOE = new HoeItem(
-            ModMaterials.Tool.IMPURE_BEDROCK, ModConfigs.hoeAttackDamage, (float) ModConfigs.hoeAttackSpeed, new Item.Properties().fireResistant().setId(key(Constants.IMPUREBEDROCK_HOE_ID))
+            ModMaterials.Tool.IMPURE_BEDROCK, config.hoeAttackDamage, (float) config.hoeAttackSpeed, new Item.Properties().fireResistant().setId(key(Constants.IMPUREBEDROCK_HOE_ID))
     );
 
     //armor