+++ CHANGELOG.md
@@ -1,1 +1,2 @@
-- Fixing error #23
+- Port to 1.21.9/10/11
+- Config file refactored